// src/generated/routes.ts
// Generated by pnpm generate. Do not edit.
import * as Procedures from "../app/procedures.js"
import { createNodeRouter } from "../runtime/server.js"

export const procedures = {
  todosByList: Procedures.todosByList,
  setTodoCompleted: Procedures.setTodoCompleted,
  createTodo: Procedures.createTodo,
  setTodoText: Procedures.setTodoText,
  deleteTodo: Procedures.deleteTodo,
  clearCompletedTodos: Procedures.clearCompletedTodos,
}

export const reactiveNeonRouter = createNodeRouter(procedures)